update Youku video link
Hi all,

We made a video to introduce the UML to YANG mapping tool we developed. It's uploaded to youtube and youku. The video links are in the readme file. 

https://github.com/OpenNetworkingFoundation/EAGLE-Open-Model-Profile-and-Tools/blob/UmlYangTools/README.md 

Please check the video and feel free to let us know if you have any comments and suggestion!


Regards

Hui and Guoying