Set default prefix if not input by the user
In order to prevent forcing the users to input the "prefix" value for
each module when using the tool, the rule has been changed as:

1. If user input the "prefix" value manually, then the generated output
prefix will be what the user inputs.
2. If the "prefix" value is kept empty (prefix "" ), then the tool will
automatically set the module name as the "prefix".
This should avoid the huge working load when there are many yang files
to generate.